에셋오디오 플레이어 에서 -> 저스트 오디오 플레이어로 바꾼후 잘동작함 굿굿
