+++ android/app/src/main/AndroidManifest.xml
@@ -11,6 +11,7 @@
         android:name="io.flutter.app.FlutterApplication"
         android:label="project_danso"
         android.requestLegacyExternalStorage="true"
+        android:usesCleartextTraffic="true"
         android:icon="@mipmap/ic_launcher">
         <provider
         android:name="androidx.core.content.FileProvider"
