+++ lib/widgets/jungganbos/jungganbo.dart
@@ -6,8 +6,6 @@ import 'package:project_danso/controllers/controllers.dart';
 import 'package:project_danso/controllers/pitch_check_controller.dart';
 import 'package:project_danso/utils/danso_function.dart';
 
-PitchCheckController pitchCheckController = Get.put(PitchCheckController());
-
 Widget jungganbo(int heightNumber, JungganboController controller,
     JungGanBo testJungGanBo, bool krState) {
   double height;
@@ -118,7 +116,7 @@ Widget changeTextColor(JungGanBo testJungGanBo, int i, int j, int heightNumber,
             ? MctSize.eighteen.getSize.sp
             : MctSize.fourteen.getSize.sp,
         color: controller.isChallenge
-            ? pitchCheckController.matchTrueFalse[i][j]
+            ? controller.matchTrueFalse[i][j]
                 ? Colors.blue
                 : Colors.red
             : MctColor.black.getMctColor),
